Refactors `recipe_parser.py`
- Refactors the bloated `recipe_parser.py` file into several supporting files.
- The existing file was starting to push 2000 lines, so it was time to break
  functionality up into several supporting files.
- There are no net logic changes in this commit.
- Outside of some constants, all entities moved were private members of the
  original file/class. This should not require much in the way of changes
  for external users (the path to `recipe_parser.py` changed slightly
- Comments now conform to the 120 character line limit, not the previous 80
  character limit. This further reduces line numbers
