+++ percy/render/parser/__init__.py
@@ -1,0 +1,1 @@
+# This package contains supporting files for the `RecipeParser` class
